Add support for modes within the application
Currently users have to type add contact or add meeting
in order to add a contact or meeting respectively.
This also applies to other commands such as edit and delete.

To improve upon this, a mode feature was implemented so that
users can be either in the contact mode or meeting mode and
when a command is run it will automatically be translated
into the respective contact or meeting command.