Implement filter by criteria
Users have to go through the entire contact and meeting list to find the
entries they want. The find function of find only covers contact name.

Let's,
modify the list function such that user can enter
attributes of the contact or meeting such as name, title, address, place
and so on, to list only the contact or meetings that matches these
attributes.

This will give user greater convenience in finding the entry that they
are looking for.

The filter keywords are case insensitive, order of the keywords does not
matter and entries matching at least one keyword will be returned.